+++ node-src/src/types/index.ts
@@ -1,3 +1,2 @@
 export * from './enums';
-// export * from './models';
-export * from './schema/schema';
+export * from './models';
