+++ node-src/src/types/models/Bid.model.ts
@@ -1,9 +1,12 @@
-import { components, paths } from '..';
+export interface BidPayload {
+  value: number;
+  itemId: string;
+  receivingDate: string;
+}
 
-export type BidPayload = components['schemas']['BidInput'];
-
-export type Bid = components['schemas']['Bid'];
-
-export type WinnerPayload = components['schemas']['WinnerInput'];
-
-export type BidPutRequestBody = paths['/api/Item/bid']['put']['requestBody'];
+export interface Bid {
+  value: number;
+  receivingDate: string;
+  user?: string;
+  createdAt?: string;
+}
