+++ node-src/src/types/models/index.ts
@@ -1,5 +1,4 @@
-// export * from './Item.model';
-// export * from './User.model';
-// export * from './Bid.model';
-// export * from './Login.model';
-// export * from './Project.model';
+export * from './Item.model';
+export * from './User.model';
+export * from './Bid.model';
+export * from './Project.model';
