--- node-src/src/types/old/enums/ItemStatus.ts
@@ -1,5 +1,0 @@
-export enum ItemStatus {
-  Undecided = 0,
-  Decided = 1,
-  Goodwill = 2
-}
