--- node-src/src/types/old/index.ts
@@ -1,2 +1,0 @@
-export * from './enums';
-export * from './models';
