--- node-src/src/types/old/models/Bid.model.ts
@@ -1,12 +1,0 @@
-export interface BidPayload {
-  value: number;
-  itemId: string;
-  receivingDate: string;
-}
-
-export interface Bid {
-  value: number;
-  receivingDate: string;
-  user?: string;
-  createdAt?: string;
-}
