--- node-src/src/types/old/models/User.model.ts
@@ -1,11 +1,0 @@
-export interface User {
-  id: string;
-  firstName: string;
-  lastName: string;
-  emailAddress: string;
-  isAdmin?: boolean;
-  ownedProjects?: string[];
-  endowmentProjects?: string[];
-  phoneNumber?: string;
-  address?: string;
-}
