--- node-src/src/types/old/models/index.ts
@@ -1,3 +1,0 @@
-export * from './Item.model';
-export * from './User.model';
-export * from './Bid.model';
