+++ node-src/src/types/openApi/models/Login.model.ts
@@ -1,3 +1,3 @@
-import { components } from '..';
+import { components } from '../schema/schema';
 
 export type Login = components['schemas']['Login'];
