+++ node-src/src/types/openApi/schema/schema.ts
